% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{RaMPP.lib}
\alias{RaMPP.lib}
\title{RaMPP.lib}
\usage{
RaMPP.lib()
}
\description{
This function runs the shiny app that provides easy and interactive access and 
visualization of the Raman spectra of 'RaMPP'. Additionally, this app allows the 
comparison of the available spectra and also with a user input spectrum. Upon mouse hovering 
on the provided graphs, it will further show which components underlie the current positions'
spectrum and whether this is a typical peak for a certain polymer.
}
\references{
TODO: Link to the Publication
}
